fix in dailyspin add and delete, forced to be at least 4

+++ app/Http/Controllers/DailySpinController.php
@@ -17,10 +17,17 @@ class DailySpinController extends Controller
 
     public function add(Request $request){
         $request->validate([
-            "entries" => "required|array|min:4",
+            "entries" => "required|array",
             "entries.*.text"=> "required|string",
             "entries.*.prize"=> "required|numeric|min:0"
         ]);
+
+        //check if the new entries lenght + the current entries length is greater or equal to 4
+        $adminId = Auth::guard('admin')->user()->getId();
+        $currentEntries = SpinWheelEntry::where('admin_id', $adminId)->get();
+        if(count($currentEntries) + count($request->input('entries')) < 4){
+            return redirect()->back()->with('error', 'You can only have 4 entries');
+        }
         
         $entries = $request->input('entries');
         $admin = Auth::guard('admin')->user();
@@ -40,6 +47,12 @@ class DailySpinController extends Controller
         $request->validate([
             "id" => "required|integer"
         ]);
+
+        $adminId = Auth::guard('admin')->user()->getId();
+        $currentEntries = SpinWheelEntry::where('admin_id', $adminId)->get();
+        if(count($currentEntries) - 1 < 4){
+            return redirect()->back()->with('error', "You can't have less than 4 entries");
+        }
 
         $entry = SpinWheelEntry::find($request->input('id'));
         $entry->delete();
